% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generator.R
\name{data_generator}
\alias{data_generator}
\title{Data Generator}
\usage{
data_generator(n, p, Delta = NULL, case = "sparse", seed = NULL)
}
\arguments{
\item{n}{The number of observations generated.}

\item{p}{The number of dimensions for the generated samples.}

\item{Delta}{Optional parameter - Provides the differential network that will be used to obtain the sample covariance matrices.}

\item{case}{Optional parameter - Selects under which case the covariance matrices are determined. Possible cases are: "sparse" - Sparse Case or "asymsparse"- Asymptotically Sparse Case. Defaults to "sparse".}

\item{seed}{Optional parameter - Allows a seed to be set for reproducibility.}
}
\value{
A list with the case used (\strong{case}), seed selected (\strong{seed_option}), two multivariate normal samples (\strong{X} and \strong{Y}), their covariance matrices (\strong{Sigma_X} and \strong{Sigma_Y}), their precision matrices (\strong{Omega_X} and \strong{Omega_Y}), the difference of precision matrices (\strong{diff_Omega}) and the target differential network (\strong{Delta}).
}
\description{
This functions generates two \eqn{n} by \eqn{p} size samples of multivariate normal
data. In doing this it also determines and provides the relevant covariance
matrices.
}
\examples{
data_generator(n = 100, p = 50, seed = 123)
data_generator(n = 10, p = 50, case = "asymsparse")
}
